+++ actions/pick-feature.js
@@ -2,13 +2,45 @@
  * @file
  * 
  * ORIGINAL SOURCE: g3w-client-plugin-editing/workflows/index.j@v4.0.0
+ * ORIGINAL SOURCE: g3w-client-plugin-editing/interactions/pickfeatures.j@v4.0.0
  * 
  * @since g3w-client-plugin-editing@v4.1.0
  */
 
-import { setAndUnsetSelectedFeaturesStyle }             from '../utils/setAndUnsetSelectedFeaturesStyle';
-import { PickFeaturesInteraction }                      from '../actions/pick-features';
-import { Step }                                         from '../g3w-step';
+import { setAndUnsetSelectedFeaturesStyle } from '../utils/setAndUnsetSelectedFeaturesStyle';
+import { Step }                             from '../g3w-step';
+
+/**
+ * @see https://openlayers.org/en/v5.3.0/apidoc/module-ol_interaction_Pointer.html
+ */
+export class PickFeaturesInteraction extends ol.interaction.Pointer {
+
+  constructor(opts = {}) {
+    let features = []; // picked features
+
+    const featuresAtPixel = ({ pixel, map } = {}) => map.getFeaturesAtPixel(pixel, {
+      layerFilter: l => opts.layer === l,
+      hitTolerance: (isMobile && isMobile.any) ? 10 : 0,
+    });
+
+    super({
+      handleDownEvent(e) {
+        features = featuresAtPixel(e);
+        return features;
+      },
+      handleUpEvent(e) {
+        if (features && features.length > 0) {
+          this.dispatchEvent({ type: 'picked', features, coordinate: e.coordinate, layer: opts.layer });
+        }
+        return true;
+      },
+      handleMoveEvent(e) {
+        e.map.getTargetElement().style.cursor = featuresAtPixel(e) ? 'pointer': '';
+      }
+    });
+  }
+
+}
 
 /**
  * ORIGINAL SOURCE: g3w-client-plugin-editing/workflows/steps/tasks/pickfeaturetask.js@v3.7.1
