+++ components/FormRelation.vue
@@ -271,11 +271,11 @@
   import { chooseFeatureFromFeatures }                    from '../utils/chooseFeatureFromFeatures';
   import { isSameBaseGeometryType }                       from '../utils/isSameBaseGeometryType';
   import { unlinkRelation }                               from '../utils/unlinkRelation';
-  import { PickFeaturesInteraction }                      from '../actions/pick-features';
+  import { PickFeaturesInteraction }                      from '../actions/pick-feature';
   import { OpenFormStep }                                 from '../actions/open-form';
   import { OpenTableStep }                                from '../actions/open-table';
   import { AddFeatureStep }                               from '../actions/add-feature';
-  import { ModifyGeometryVertexStep }                     from '../actions/modify-geometry-vertex';
+  import { ModifyGeometryVertexStep }                     from '../actions/modify-geometry';
   import { MoveFeatureStep }                              from '../actions/move-feature';
 
   const { ProjectsRegistry }            = g3wsdk.core.project;
